+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/reference/GenomeVariantNormalizer.java
@@ -53,7 +53,7 @@ public final class GenomeVariantNormalizer {
 
 		// Insert the ALT bases at the position indicated by txPos.
 		int pos = txPos.getPos();
-		StringBuilder builder = new StringBuilder(transcript.getSequence());
+		StringBuilder builder = new StringBuilder(transcript.getTrimmedSequence());
 		builder.insert(pos, change.getAlt());
 
 		// Execute algorithm and compute the shift.
@@ -106,7 +106,7 @@ public final class GenomeVariantNormalizer {
 		// Shift the deletion to the 3' (right) end of the transcript.
 		int pos = txPos.getPos();
 		final int LEN = change.getRef().length(); // length of the deletion
-		final String seq = transcript.getSequence();
+		final String seq = transcript.getTrimmedSequence();
 		int shift = 0;
 
 		while ((pos + LEN < seq.length()) && (seq.charAt(pos) == seq.charAt(pos + LEN))) {