+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/reference/TranscriptInterval.java
@@ -91,7 +91,7 @@ public final class TranscriptInterval {
 	 */
 	@Override
 	public String toString() {
-		return StringUtil.concatenate(transcript.getAccession(), ":n.", beginPos + 1, "-", endPos);
+		return StringUtil.concatenate(transcript.getAccession(), ":n.", beginPos + 1, "_", endPos);
 	}
 
 	/*